requirements.txt
# Core
python>=3.10
ollama>=0.1.0

# Audio (Modern alternatives)
sounddevice>=0.4.6
scipy>=1.11.0
edge-tts>=6.1.9
vosk>=0.3.45  # Offline STT
pygame>=2.5.0

# Vision & OCR
opencv-python>=4.8.0
pytesseract>=0.3.10
pillow>=10.0.0

# Utilities
pyautogui>=0.9.54
psutil>=5.9.5
schedule>=1.2.0
python-dotenv>=1.0.0
PyYAML>=6.0.1

# APIs
duckduckgo-search>=3.9.0
wikipedia>=1.4.0

# Machine Learning & Search
scikit-learn>=1.3.0
chromadb>=0.4.0
trafilatura>=1.6.3